created pet profile page. Attached it to the pet cards. Added images into img folder until back end adds them then I can call them from the DB. Modified dropdowns and search as needed for the pet profile page. Added login/logout rules to navbar for users. Its currently commented out but once we have the login in working then the nav will update based on user status. I will also have the favories slider change to available pets if logged out just need to connect with backend team. Moved home page components into home folder and arranged components into their folders by page.
