--- src/components/PetComponents/PetData/DogBreeds.tsx
@@ -1,164 +1,0 @@
-const DogBreeds: string[] = [
-  "Mix Breeds",
-  "American Eskimo",
-  "American Hairless Terrier",
-  "Anatolian Shepherd",
-  "Australian Cattle",
-  "Australian Shepherd",
-  "Basenji",
-  "Basset Hound",
-  "Beagle",
-  "Bearded Collie",
-  "Belgian Malinois",
-  "Belgian Sheepdog",
-  "Belgian Tervuren",
-  "Bernese Mountain",
-  "Bichon Frise",
-  "Black and Tan Coonhound",
-  "Black Russian Terrier",
-  "Bloodhound",
-  "Border Collie",
-  "Border Terrier",
-  "Borzoi",
-  "Boston Terrier",
-  "Boxer",
-  "Brittany",
-  "Brussels Griffon",
-  "Bull Terrier",
-  "Bulldog (English Bulldog, French Bulldog)",
-  "Bullmastiff",
-  "Cairn Terrier",
-  "Cardigan Welsh Corgi",
-  "Cavalier King Charles Spaniel",
-  "Chesapeake Bay Retriever",
-  "Chihuahua",
-  "Chinese Shar-Pei",
-  "Chow Chow",
-  "Clumber Spaniel",
-  "Cocker Spaniel",
-  "Collie",
-  "Curly-Coated Retriever",
-  "Dachshund",
-  "Dalmatian",
-  "Dandie Dinmont Terrier",
-  "Doberman Pinscher",
-  "Dogue de Bordeaux",
-  "English Cocker Spaniel",
-  "English Foxhound",
-  "English Setter",
-  "English Springer Spaniel",
-  "English Toy Spaniel",
-  "Entlebucher Mountain",
-  "Finnish Lapphund",
-  "Finnish Spitz",
-  "Flat-Coated Retriever",
-  "French Bulldog",
-  "German Pinscher",
-  "German Shepherd",
-  "German Shorthaired Pointer",
-  "German Wirehaired Pointer",
-  "Giant Schnauzer",
-  "Glen of Imaal Terrier",
-  "Golden Retriever",
-  "Gordon Setter",
-  "Great Dane",
-  "Great Pyrenees",
-  "Greater Swiss Mountain",
-  "Greyhound",
-  "Havanese",
-  "Irish Red and White Setter",
-  "Irish Setter",
-  "Irish Terrier",
-  "Irish Water Spaniel",
-  "Irish Wolfhound",
-  "Italian Greyhound",
-  "Japanese Chin",
-  "Japanese Spitz",
-  "Keeshond",
-  "Kerry Blue Terrier",
-  "Komondor",
-  "Kuvasz",
-  "Labrador Retriever",
-  "Lakeland Terrier",
-  "Leonberger",
-  "Lhasa Apso",
-  "Lowchen",
-  "Maltese",
-  "Manchester Terrier",
-  "Mastiff",
-  "Miniature American Shepherd",
-  "Miniature Bull Terrier",
-  "Miniature Pinscher",
-  "Miniature Schnauzer",
-  "Neapolitan Mastiff",
-  "Newfoundland",
-  "Norfolk Terrier",
-  "Norwegian Buhund",
-  "Norwegian Elkhound",
-  "Norwegian Lundehund",
-  "Norwich Terrier",
-  "Nova Scotia Duck Tolling Retriever",
-  "Old English Sheepdog",
-  "Otterhound",
-  "Papillon",
-  "Parson Russell Terrier",
-  "Pekingese",
-  "Pembroke Welsh Corgi",
-  "Petit Basset Griffon Vendeen",
-  "Pharaoh Hound",
-  "Plott",
-  "Pointer",
-  "Polish Lowland Sheepdog",
-  "Pomeranian",
-  "Poodle (Standard, Miniature, Toy)",
-  "Portuguese Podengo Pequeno",
-  "Portuguese Water Dog",
-  "Pug",
-  "Puli",
-  "Pumi",
-  "Pyrenean Shepherd",
-  "Rat Terrier",
-  "Redbone Coonhound",
-  "Rhodesian Ridgeback",
-  "Rottweiler",
-  "Russell Terrier",
-  "Saluki",
-  "Samoyed",
-  "Schipperke",
-  "Scottish Deerhound",
-  "Scottish Terrier",
-  "Sealyham Terrier",
-  "Shetland Sheepdog",
-  "Shiba Inu",
-  "Shih Tzu",
-  "Siberian Husky",
-  "Silky Terrier",
-  "Skye Terrier",
-  "Sloughi",
-  "Small Munsterlander Pointer",
-  "Soft-Coated Wheaten Terrier",
-  "Spanish Water Dog",
-  "Spinone Italiano",
-  "Staffordshire Bull Terrier",
-  "Standard Schnauzer",
-  "Sussex Spaniel",
-  "Swedish Vallhund",
-  "Tibetan Mastiff",
-  "Tibetan Spaniel",
-  "Tibetan Terrier",
-  "Toy Fox Terrier",
-  "Treeing Walker Coonhound",
-  "Vizsla",
-  "Weimaraner",
-  "Welsh Springer Spaniel",
-  "Welsh Terrier",
-  "West Highland White Terrier",
-  "Whippet",
-  "Wire Fox Terrier",
-  "Wirehaired Pointing Griffon",
-  "Wirehaired Vizsla",
-  "Xoloitzcuintli",
-  "Yorkshire Terrier",
-];
-
-export default DogBreeds;
