--- src/components/PetComponents/PetData/GuineaPigBreeds.tsx
@@ -1,15 +1,0 @@
-const GuineaPigBreeds = [
-  "American",
-  "Abyssinian",
-  "Peruvian",
-  "Silkie",
-  "Teddy",
-  "Texel",
-  "Coronet",
-  "Crested",
-  "Sheltie",
-  "Skinny",
-  "Baldwin",
-];
-
-export default GuineaPigBreeds;
