+++ src/components/SearchComponents/SearchInput.tsx
@@ -166,7 +166,7 @@ const SearchInput: React.FC<SearchInputProps> = ({
   }, [keyword, location, type, sex, age, breed, favorite]);
 
   return (
-    <Box sx={{ mx: "1rem" }}>
+    <Box sx={{ mx: "1rem", mb: "3rem" }}>
       <Stack spacing={2} sx={{ ml: "1rem", pt: "2rem" }}>
         <TextField
           variant="outlined"
@@ -210,10 +210,8 @@ const SearchInput: React.FC<SearchInputProps> = ({
         New Search
       </Button>
 
-      <Card
-        sx={{ marginTop: "2rem", py: 5, px: 3, backgroundColor: "#0E2728" }}
-      >
-        <Stack gap={3}>
+      <Card sx={{ mt: "2rem", py: 5, px: 3, backgroundColor: "#0E2728" }}>
+        <Stack gap={4}>
           <FilterDropdowns
             type={type}
             breed={breed}
