+++ src/index.css
@@ -1,3 +1,3 @@
-:root {
+html {
   background-color: #f4f2ea;
 }
