+++ src/util/PetData/CatBreeds.tsx
@@ -1,0 +1,42 @@
+const CatBreeds: string[] = [
+  "Abyssinian",
+  "American Bobtail",
+  "American Curl",
+  "American Shorthair",
+  "American Wirehair",
+  "Balinese",
+  "Bengal",
+  "Birman",
+  "Bombay",
+  "British Shorthair",
+  "Burmese",
+  "Chartreux",
+  "Cornish Rex",
+  "Devon Rex",
+  "Domestic Longhair",
+  "Domestic Shorthair",
+  "Egyptian Mau",
+  "Exotic Shorthair",
+  "Himalayan",
+  "Japanese Bobtail",
+  "Maine Coon",
+  "Manx",
+  "Norwegian Forest Cat",
+  "Oriental",
+  "Persian",
+  "Ragamuffin",
+  "Ragdoll",
+  "Russian Blue",
+  "Scottish Fold",
+  "Siamese",
+  "Siberian",
+  "Singapura",
+  "Somali",
+  "Sphynx",
+  "Tabby",
+  "Tonkinese",
+  "Turkish Angora",
+  "Turkish Van",
+];
+
+export default CatBreeds;
