+++ src/util/PetData/GuineaPigBreeds.tsx
@@ -1,0 +1,15 @@
+const GuineaPigBreeds = [
+  "American",
+  "Abyssinian",
+  "Peruvian",
+  "Silkie",
+  "Teddy",
+  "Texel",
+  "Coronet",
+  "Crested",
+  "Sheltie",
+  "Skinny",
+  "Baldwin",
+];
+
+export default GuineaPigBreeds;
