+++ src/util/PetData/HamsterBreeds.tsx
@@ -1,0 +1,14 @@
+const HamsterBreeds = [
+  "Syrian Hamster",
+  "Dwarf Campbell Russian Hamster",
+  "Dwarf Winter White Russian Hamster",
+  "Roborovski Hamster",
+  "Chinese Hamster",
+  "Teddy Bear Hamster",
+  "Golden Hamster",
+  "Panda Hamster",
+  "European Hamster",
+  "Robby Hamster",
+];
+
+export default HamsterBreeds;
