+++ src/util/PetData/RabbitBreeds.tsx
@@ -1,0 +1,14 @@
+const RabbitBreeds: string[] = [
+  "Dutch Rabbit",
+  "Mini Lop",
+  "Holland Lop",
+  "Netherland Dwarf",
+  "Lionhead Rabbit",
+  "Flemish Giant",
+  "English Angora",
+  "French Angora",
+  "Rex Rabbit",
+  "Mini Rex",
+];
+
+export default RabbitBreeds;
